fix: remove generated components script
