+++ .gitignore
@@ -6,7 +6,6 @@ tmp
 out-tsc
 storybook-static
 storybook-styles/*.css
-.storybook/generated
 # Only exists if Bazel was run
 bazel-out
