--- .storybook/generated-types.d.ts
@@ -1,7 +1,0 @@
-// Type declaration for the generated JSON file.
-// The file is created by scripts/generate-deprecated-components.js before Storybook starts.
-// This declaration prevents TS errors when the file doesn't exist (e.g. in CI build).
-declare module "*/generated/deprecated-components.json" {
-  const value: Record<string, string>;
-  export default value;
-}
